% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Lauch the shiny app}
\usage{
launch_app(x, ...)
}
\arguments{
\item{x}{your LC credential}

\item{...}{pass additional arguments to shiny::runApp}
}
\description{
Opens a shiny app capable of retrieving Lending Club account data and to
 execute transactions.
}
