% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClubAccess.R
\docType{package}
\name{LendingClubAccess}
\alias{LendingClubAccess}
\alias{LendingClubAccess-package}
\title{Shiny app for woking with a LendingClub account}
\description{
Provides an alternative method for many of the common actions taken on the
Lending Club website.  Investors are able to access their account summary,
transfer money in or out of their account, and most importantly buy and sell
notes.
}
